==> scripts/04_rates_24h.Rmd <==
---
title: "Relationships between baseline characteristics and average mearement rates"
author: "Seraphina Shi"
output:
  html_document:
    toc: true
    toc_float: true
---


```{r load_lib, include = FALSE, warning=FALSE, message=FALSE, echo=FALSE} 
# remotes::install_github("tlverse/sl3@devel")
# remotes::install_github("tlverse/tmle3")

library(dplyr)
library(tidyr)
library(data.table)
library(ggplot2)
library(here)
library(future)
library(gridExtra)
library(corrplot)
library(tmle3)
library(sl3)
library(table1)
library(gtsummary)
library(knitr)

```

```{r setup, include = FALSE}
plotFolder <- here("results","images", "04_24h")
if(!file.exists(plotFolder)) dir.create(plotFolder,recursive=TRUE)

rdataFolder <- here("data", "rdata")

knitr::opts_chunk$set(
  cache=FALSE, autodep=FALSE, warning=FALSE, message=FALSE, echo=FALSE,
  results = 'markup', dev='png', dpi=150, fig.align = "center", fig.path=paste0(plotFolder, "/"),
  cache.path=".cache/",
  duplicate.label="allow"
)

set.seed(123)
```

# Prepare

```{r}
load(here(rdataFolder, "02_2_varFusion.RData"))

PreprocessDataFolder <- here("data", "DataPreprocessing")

df <- read.csv(here(PreprocessDataFolder,"02_3_data_all_blocks.csv"), row.names = 1, header = T)  
df_imp <- read.csv(here(PreprocessDataFolder,"02_3_data_imputed_all_blocks.csv"), row.names = 1, header = T)
```

```{r}
# discharged within the first 24 hours
df_tmp <- df[df$half_day <= 2, ]
discharge_in24_id <- df_tmp$ICUSTAY_ID[df_tmp$Y_t == 2]

df <- df[! df$ICUSTAY_ID %in% discharge_in24_id, ]
df_imp <- df_imp[! df_imp$ICUSTAY_ID %in% discharge_in24_id, ]
```

```{r}
data_ids <- read.csv(here(PreprocessDataFolder,"data_ids.csv"), row.names = 1, header = T)  

data_ids <- data_ids[data_ids$ICUSTAY_ID %in% df$ICUSTAY_ID, ]
length(unique(data_ids$SUBJECT_ID))
length(unique(data_ids$ICUSTAY_ID))
```

```{r, results='asis'}
df_24h <- df_imp %>% 
  filter(half_day <= 2) %>% 
  group_by(ICUSTAY_ID) %>% 
  mutate(outcome = max(Y_t),
         less_24h = as.numeric(outcome > 0)) %>%
  dplyr::select(ICUSTAY_ID, half_day, less_24h, outcome, Y_t, Y_t_plus_1, everything()) %>%
  ungroup() %>%
  mutate(outcome = as.character(outcome),
         Y_t = as.character(Y_t))

# View(df_24h %>% dplyr::select(ICUSTAY_ID, half_day, less_24h, outcome, Y_t, Y_t_plus_1))

df_24h <- df_24h %>%
  dplyr::select(-c(Y_t, Y_t_plus_1))

df_24h_1 <- df_24h %>%
  dplyr::select(ICUSTAY_ID, outcome, less_24h) %>%
  unique()

df_24h_2 <- df_24h %>% 
  dplyr::select(unname(c("ICUSTAY_ID", 
                         var_name_list$num_vitals, 
                         var_name_list$hours_missing_vitals, 
                         var_name_list$num_labs))) %>%
  group_by(ICUSTAY_ID) %>% 
  summarise(across(everything(), sum))

df_24h_3 <- df_24h %>% 
  dplyr::select(unname(c("ICUSTAY_ID", 
                         var_name_list$num_baselines, 
                         var_name_list$baselines, 
                         var_name_list$vitals, 
                         var_name_list$labs))) %>%
  group_by(ICUSTAY_ID) %>% 
  summarise(across(everything(), mean))

df_24h <- merge(df_24h_1, df_24h_2, by="ICUSTAY_ID") %>%
  merge(df_24h_3, by="ICUSTAY_ID")
rm(df_24h_1, df_24h_2)
```



Proportion of whether died within the 24h:
```{r}
table1(~outcome, data = df_24h)
```

```{r}
# standardize the measurement rates for patients that died within the first 24 hours
library(lubridate)
ADMISSIONS <- read.csv(here("data","raw_data","mimic-iii-clinical-database-1.4", "ADMISSIONS.csv")) %>% 
  select(SUBJECT_ID, HADM_ID, DEATHTIME)
PATIENTS <- read.csv(here("data","raw_data","mimic-iii-clinical-database-1.4", "PATIENTS.csv")) %>% 
  select(SUBJECT_ID, DOD_SSN)
ICUSTAYS <- read.csv(here("data","raw_data","mimic-iii-clinical-database-1.4", "ICUSTAYS.csv")) %>% 
  select(SUBJECT_ID, HADM_ID, ICUSTAY_ID, INTIME, OUTTIME)

df_Y <- ADMISSIONS %>% 
  merge(PATIENTS, by = "SUBJECT_ID", all = T) %>%
  merge(ICUSTAYS, by = c("SUBJECT_ID", "HADM_ID"), all = T) %>%
  unique()

df_Y$DOD_SSN[is.na(df_Y$DOD_SSN)] = df_Y$DEATHTIME[is.na(df_Y$DOD_SSN)]
df_Y$DEATHTIME[df_Y$DEATHTIME == ""] =  df_Y$DOD_SSN[df_Y$DEATHTIME == ""]

df_Y <- df_Y %>% 
  mutate(INTIME =  ymd_hms(INTIME),
         OUTTIME = ymd_hms(OUTTIME),
         DEATHTIME = ymd_hms(DEATHTIME),
         DISCHARGE_AFTER_ICUS_IN_hours = (as.numeric(difftime(OUTTIME, INTIME))/(60*60)),
         DEATH_AFTER_ICUS_IN_hours =  (as.numeric(difftime(DEATHTIME, INTIME))/60)
         ) 

df_Y <- df_Y %>%  rowwise() %>%
  mutate(ICU_length_hours = min(c(DISCHARGE_AFTER_ICUS_IN_hours, DEATH_AFTER_ICUS_IN_hours), na.rm = T)) %>%
  select(ICUSTAY_ID, ICU_length_hours) 


df_24h <- merge(df_24h, df_Y, by = "ICUSTAY_ID")

df_24h_1 <- df_24h %>% filter(outcome==1)
for (var in c(var_name_list$hours_missing_vitals, var_name_list$num_vitals, var_name_list$num_labs)) {
  df_24h_1[var] = 24 * df_24h_1[var] /  df_24h_1$ICU_length_hours
}

df_24h <- rbind(df_24h_1, df_24h %>% filter(outcome==0))
```


```{r get_severity_score}
sofa <- read.csv(here("data", "SeverityScores", "sofa.csv")) 
sapsii <- read.csv(here("data", "SeverityScores", "sapsii.csv")) 

df_24h <- df_24h %>% 
  merge(sofa %>% 
          dplyr::select(icustay_id, sofa) %>% 
          rename(ICUSTAY_ID = icustay_id), 
        by="ICUSTAY_ID",
        all.x = T) %>%
  merge(sapsii %>% 
          dplyr::select(icustay_id, sapsii) %>% 
          rename(ICUSTAY_ID = icustay_id), 
        by="ICUSTAY_ID",
        all.x = T)  

```


```{r}
dim(df_24h)
length(unique(df_24h$ICUSTAY_ID))
```

```{r prepare_function}
source(here("scripts", "04_rates_24h_functions.R"))
```




# Correlation plots
## Baseline, vital sign measurement rates, severity scores
```{r corr_base_n_vital, fig.height = 8, fig.width = 8}
cor_mat <- cor(df_24h %>% select(var_name_list$baselines,  c("sofa", "sapsii"), var_name_list$num_vitals))
n_base <- length(var_name_list$baselines) + 2
n_y <- length(var_name_list$num_vitals)
corrplot(cor_mat[1:n_base,(n_base-1):(n_base+n_y)], method="color")
```

## Baseline, vital sign missing rates, severity scores
```{r corr_base_h_m_vital, fig.height = 8, fig.width = 8}
cor_mat <- cor(df_24h %>% select(var_name_list$baselines,  c("sofa", "sapsii"), var_name_list$hours_missing_vitals))
n_base <- length(var_name_list$baselines) + 2
n_y <- length(var_name_list$hours_missing_vitals)
corrplot(cor_mat[1:n_base,(n_base-1):(n_base+n_y)], method="color") #mar = c(0, 0, 0, 0)
```

## Baseline, lab testing rates, severity scores
```{r corr_base_n_lab, fig.height = 8, fig.width = 8}
cor_mat <- cor(df_24h %>% select(var_name_list$baselines,  c("sofa", "sapsii"), var_name_list$num_labs))
n_base <- length(var_name_list$baselines) + 2
n_y <- length(var_name_list$num_labs)
corrplot(cor_mat[1:n_base,(n_base-1):(n_base+n_y)], method="color") #mar = c(0, 0, 0, 0)
```


# AGE vs Rates
## 00. Age distribution
```{r age_hist}
cat("Summary of age distribution: \n")
hist(df_24h$AGE)
summary(df_24h$AGE)
age_c = cut(df_24h$AGE,
            breaks = c(18,30,45,65,101),
            labels = c("18-30","31-45","46-65",">65"))
df_24h <- df_24h %>% mutate(age_cat = age_c)
rm(age_c)

df_24h_all <- df_24h

cat("Summary of age groups: \n")
table1(~age_cat, data = df_24h)

table1(~sofa+sapsii|age_cat, data = df_24h)
```

```{r age_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = age_cat, y=sofa, col=age_cat)) +
        theme(legend.position = "none") +
        labs(x = "Age group", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = age_cat, y=sapsii, col=age_cat)) +
        theme(legend.position = "none") +
        labs(x = "Age group", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "age_cat"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```



```{r, echo=FALSE, results='asis'}
# baselines = var_name_list$baselines[var_name_list$baselines != "AGE"]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# idx <- 1:length(y_vars)
# 
# glmfit_0_rlt_list <- list()
# glmfit_sofa_rlt_list <- list()
# glmfit_saps_rlt_list <- list()
# tmle_fit_base_rlt_list <- list()
# tmle_fit_base_sofa_rlt_list <- list()
# tmle_fit_base_saps_rlt_list <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
#   
#   source(' 04_rates_24h_template.R')
#   
#   glmfit_0_rlt_list[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("haha\n\n")
# }
# 
# # save.image(here(rdataFolder, '04_24h_age.RData'))
# save.image(here('04_24h_age.RData'))
```

## 0. TMLE estimations plots 

```{r}
load(here('data', 'rdata', '04_24h_age.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_age_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list) %>% as.data.frame()

tmle_fit_base_sofa_rlt_age_all_tsm <- tmle_fit_base_sofa_rlt_age_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital", 
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_age_all_tsm$group = factor(tmle_fit_base_sofa_rlt_age_all_tsm$group, levels = c('18-30', '31-45', '46-65', '>65'))

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)

```

```{r fig.width=7, fig.height=2.3}
plotFolder <- here("results","images", "04_24h")

rlts_s <- tmle_fit_base_sofa_rlt_age_all_tsm[tmle_fit_base_sofa_rlt_age_all_tsm$Y_name %in% c("BP", "HR", "TempC"), ]
g <- make_box_plots_tmle_rlts(rlts_s, 'n_vital') + labs(title = "Vital Sign Measurement Frequencies") + theme(axis.text.x = element_text(angle = 0, vjust = 0, hjust=0.5)) 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n_small.png")), g, width = 7.5, height = 2.3, dpi = 1200)

g <- make_box_plots_tmle_rlts(rlts_s, 'h_m_vital') + labs(title = "Vital Sign Missing Hours") + theme(axis.text.x = element_text(angle = 0, vjust = 0, hjust=0.5)) 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m_small.png")), g, width = 7.5, height = 2.3, dpi = 1200)

rlts_s <- tmle_fit_base_sofa_rlt_age_all_tsm[tmle_fit_base_sofa_rlt_age_all_tsm$Y_name %in% c("Lac", "lab_grp1", "lab_grp6"), ]
g <- make_box_plots_tmle_rlts(rlts_s, 'n_lab') + labs(title = "Laboratory Testing Frequencies")   + theme(axis.text.x = element_text(angle = 0, vjust = 0, hjust=0.5)) 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n_small.png")), g, width = 7.5, height = 2.3, dpi = 1200)
```




```{r fig.width=16, fig.height=3}
# glmfit_0_rlt_age_all <- do.call("rbind", glmfit_0_rlt_list) %>% as.data.frame()
# glmfit_0_rlt_age_all$param <- rownames(glmfit_0_rlt_age_all)
# 
# 
# tmle_fit_base_sofa_rlt_age_all_tsm <- glmfit_0_rlt_age_all %>%
#   mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
#          Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital", 
#                          ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
#                                 ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
#          Y_name = sub("^(n_|h_m_)", "", Y))
# 
# tmle_fit_base_sofa_rlt_age_all_tsm$group = factor(tmle_fit_base_sofa_rlt_age_all_tsm$group, levels = c('18-30', '31-45', '46-65', '>65'))
# 
# g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_all_tsm, 'h_m_vital') +
#   labs(title = "Vital Sign Missing Hours") 
# g
# ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)
# 
# 
# g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_all_tsm, 'n_vital') +
#   labs(title = "Vital Sign Measurement Frequencies") 
# g
# ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)
# 
# 
# g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_all_tsm, 'n_lab') +
#   labs(title = "Laboratory Testing Frequencies") 
# g
# ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)

```



```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))
  
  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) + 
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)
  
  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list[[i]])
  print(formatted_table)

  
  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list[[i]])
  print(formatted_table)
  
}

```

# Ethnicity vs Rates
## Ethnicity distribution
```{r}
df_baseline <- read.csv(here("data", "DataPreprocessing", "df_baseline.csv"))  %>% dplyr::select(-X)
df_24h <- merge(df_24h, df_baseline %>% select(ICUSTAY_ID, ETHNICITY), by="ICUSTAY_ID", all.x = T )

df_24h$ETHNICITY[is.na(df_24h$ETHNICITY)] = "Missing"

df_24h$ethnicity = factor(df_24h$ETHNICITY , levels = c("WHITE", "ASIAN", "BLACK", "HISPANIC", "Other", "Missing"))
df_24h_all$ethnicity <- df_24h$ethnicity

cat("Summary of ethnicities: \n")
table1(~ethnicity, data = df_24h)

table1(~sofa+sapsii|ethnicity, data = df_24h)
```

Ethnicity distribution among patients survived or died in ICU within the first 24 hours
```{r}
table1(~ethnicity|outcome, data = df_24h)
```


```{r ethnicity_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = ethnicity, y=sofa, col=ethnicity)) +
        theme(legend.position = "none") +
        labs(x = "Age group", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = ethnicity, y=sapsii, col=ethnicity)) +
        theme(legend.position = "none") +
        labs(x = "Age group", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "ethnicity"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)

```


```{r}
# x = "ethnicity"
# baselines = var_name_list$baselines[! var_name_list$baselines %in% c("ETHNICITY_W", "ETHNICITY_B", "ETHNICITY_A", "ETHNICITY_H", "ETHNICITY_O")]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# idx <- 1:length(y_vars)
# 
# glmfit_0_rlt_list_eth <- list()
# glmfit_sofa_rlt_list_eth  <- list()
# glmfit_saps_rlt_list_eth  <- list()
# tmle_fit_base_rlt_list_eth  <- list()
# tmle_fit_base_sofa_rlt_list_eth  <- list()
# tmle_fit_base_saps_rlt_list_eth  <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_eth[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_eth[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_eth[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_eth[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_eth[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_eth[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("haha\n\n")
# }
# save.image(here('04_24h_eth.RData'))

```


## 0. TMLE estimations plots 

```{r}
load(here('data', 'rdata', '04_24h_eth.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_eth_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_eth) %>% as.data.frame()

tmle_fit_base_sofa_rlt_eth_all_tsm <- tmle_fit_base_sofa_rlt_eth_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital", 
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_eth_all_tsm$group = factor(tmle_fit_base_sofa_rlt_eth_all_tsm$group, levels = c("WHITE", "ASIAN", "BLACK", "HISPANIC", "Other"))

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_eth_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_eth_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_eth_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```


```{r fig.width=10, fig.height=2.5}
rlts_s <- tmle_fit_base_sofa_rlt_eth_all_tsm[tmle_fit_base_sofa_rlt_eth_all_tsm$Y_name %in% c("BP", "HR", "RR", "SpO2", "TempC"), ]
g <- make_box_plots_tmle_rlts(rlts_s, 'n_vital') + labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n_small.png")), g, width = 10.5, height = 2.5, dpi = 1200)

rlts_s <- tmle_fit_base_sofa_rlt_eth_all_tsm[tmle_fit_base_sofa_rlt_eth_all_tsm$Y_name %in% c("HCT", "lab_grp2", "lab_grp4"), ]
g <- make_box_plots_tmle_rlts(rlts_s, 'n_lab') + labs(title = "Laboratory Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n_small.png")), g, width = 7, height = 2.5, dpi = 1200)
```


```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))
  
  # cat("Summary of", y,  "w.r.t.", x, "groups: \n")  
  # df_24h$ethnicity[is.na(df_24h$ethnicity)] = "Missing"
  # df_24h$ethnicity = factor(df_24h$ethnicity , levels = c("WHITE", "ASIAN", "BLACK", "HISPANIC", "Other", "Missing"))
  # 
  # table_result <- table1(as.formula(paste0("~ ", y, "| ethnicity")), data = df_24h)
  # cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) + 
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)
  
  
  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_eth[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_eth[[i]])
  print(formatted_table)

  
  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_eth[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_eth[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_eth[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_eth[[i]])
  print(formatted_table)
  
}

```


# Gender vs Rates
```{r gender_hist}
df_baseline <- read.csv(here("data", "DataPreprocessing", "df_baseline.csv"), row.names = 1)
df_24h <- merge(df_24h, df_baseline %>% select(ICUSTAY_ID, GENDER), by="ICUSTAY_ID", all.x = T )
df_24h$gender = ifelse(df_24h$GENDER == "F", "Female", "Male")

df_24h$gender = factor(df_24h$gender)
df_24h_all$gender <- df_24h$gender

cat("Summary of gender: \n")
table1(~gender, data = df_24h)

table1(~sofa+sapsii|gender, data = df_24h)
```

```{r gender_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = gender, y=sofa, col=gender)) +
        theme(legend.position = "none") +
        labs(x = "Gender", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = gender, y=sapsii, col=gender)) +
        theme(legend.position = "none") +
        labs(x = "Gender", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "gender"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)
```

```{r, echo=FALSE, results='asis'}
# x = 'GENDER_F'
# baselines = var_name_list$baselines[! var_name_list$baselines %in% c("GENDER_F")]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# 
# glmfit_0_rlt_list_gender <- list()
# glmfit_sofa_rlt_list_gender  <- list()
# glmfit_saps_rlt_list_gender  <- list()
# tmle_fit_base_rlt_list_gender  <- list()
# tmle_fit_base_sofa_rlt_list_gender  <- list()
# tmle_fit_base_saps_rlt_list_gender  <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_gender[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_gender[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_gender[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_gender[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_gender[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_gender[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("\n haha\n\n")
# }
# save.image(here('04_24h_gender.RData'))

```



## 0. TMLE estimations plots
```{r}
load(here('data', 'rdata', '04_24h_gender.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```



```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_gender_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_gender) %>% as.data.frame()

tmle_fit_base_sofa_rlt_gender_all_tsm <- tmle_fit_base_sofa_rlt_gender_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_gender_all_tsm$group[tmle_fit_base_sofa_rlt_gender_all_tsm$group == 1] = 'Female'
tmle_fit_base_sofa_rlt_gender_all_tsm$group[tmle_fit_base_sofa_rlt_gender_all_tsm$group == 0] = 'Male'

tmle_fit_base_sofa_rlt_gender_all_tsm$group = factor(tmle_fit_base_sofa_rlt_gender_all_tsm$group)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_gender_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_gender_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_gender_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```


```{r fig.width=7, fig.height=2.5}
plotFolder <- here("results","images", "04_24h")

rlts_s <- tmle_fit_base_sofa_rlt_gender_all_tsm[tmle_fit_base_sofa_rlt_gender_all_tsm$Y_name %in% c("BP", "HR", "TempC"), ]
g <- make_box_plots_tmle_rlts(rlts_s, 'n_vital') + labs(title = "Vital Sign Measurement Frequencies") + theme(axis.text.x = element_text(angle = 0, vjust = 0, hjust=0.5)) 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n_small.png")), g, width = 7.5, height = 2.3, dpi = 1200)

g <- make_box_plots_tmle_rlts(rlts_s, 'h_m_vital') + labs(title = "Vital Sign Missing Hours") + theme(axis.text.x = element_text(angle = 0, vjust = 0, hjust=0.5)) 
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m_small.png")), g, width = 7.5, height = 2.3, dpi = 1200)

```


```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))

  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) +
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)


  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_gender[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_gender[[i]])
  print(formatted_table)


  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_gender[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_gender[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_gender[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_gender[[i]])
  print(formatted_table)

}

```





# Insurance vs Rates
```{r insurance_hist}
df_baseline <- read.csv(here("data", "DataPreprocessing", "df_baseline.csv"), row.names = 1)
df_24h <- merge(df_24h, df_baseline %>% select(ICUSTAY_ID, INSURANCE), by="ICUSTAY_ID", all.x = T )

df_24h$insurance = factor(df_24h$INSURANCE, levels = c('not_insuranced', 'insuranced'))
df_24h_all$insurance <- df_24h$insurance

cat("Summary of insurance: \n")
table1(~insurance, data = df_24h)

table1(~sofa+sapsii|insurance, data = df_24h)
```

```{r insurance_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = insurance, y=sofa, col=insurance)) +
        theme(legend.position = "none") +
        labs(x = "Insurance", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = insurance, y=sapsii, col=insurance)) +
        theme(legend.position = "none") +
        labs(x = "Insurance", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```

```{r, fig.height=3, fig.width=16}
x = "insurance"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)
```

```{r, echo=FALSE, results='asis'}
# x = 'INSURANCE_Y'
# baselines = var_name_list$baselines[! var_name_list$baselines %in% c("INSURANCE_Y")]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# 
# glmfit_0_rlt_list_insurance <- list()
# glmfit_sofa_rlt_list_insurance  <- list()
# glmfit_saps_rlt_list_insurance  <- list()
# tmle_fit_base_rlt_list_insurance  <- list()
# tmle_fit_base_sofa_rlt_list_insurance  <- list()
# tmle_fit_base_saps_rlt_list_insurance  <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_insurance[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_insurance[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_insurance[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_insurance[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_insurance[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_insurance[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("\n haha\n\n")
# }
# save.image(here('04_24h_insurance.RData'))

```



## 0. TMLE estimations plots
```{r}
load(here('data', 'rdata', '04_24h_insurance.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_insurance_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_insurance) %>% as.data.frame()

tmle_fit_base_sofa_rlt_insurance_all_tsm <- tmle_fit_base_sofa_rlt_insurance_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_insurance_all_tsm$group[tmle_fit_base_sofa_rlt_insurance_all_tsm$group == 1] = 'insuranced'
tmle_fit_base_sofa_rlt_insurance_all_tsm$group[tmle_fit_base_sofa_rlt_insurance_all_tsm$group == 0] = 'not_insuranced'

tmle_fit_base_sofa_rlt_insurance_all_tsm$group = factor(tmle_fit_base_sofa_rlt_insurance_all_tsm$group, levels = c('not_insuranced', 'insuranced'))

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_insurance_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_insurance_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_insurance_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```

```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))

  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) +
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)


  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_insurance[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_insurance[[i]])
  print(formatted_table)


  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_insurance[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_insurance[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_insurance[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_insurance[[i]])
  print(formatted_table)

}

```


# Language vs Rates
```{r LANGUAGE_hist}
df_baseline <- read.csv(here("data", "DataPreprocessing", "df_baseline.csv"), row.names = 1)
df_24h <- merge(df_24h, df_baseline %>% select(ICUSTAY_ID, LANGUAGE), by="ICUSTAY_ID", all.x = T )

df_24h$language <- ifelse(is.na(df_24h$LANGUAGE), "Missing", df_24h$LANGUAGE)

df_24h$language = factor(df_24h$language, levels = c("ENGL", "Other", "Missing"))
df_24h_all$language <- df_24h$language


cat("Summary of language: \n")
table1(~language, data = df_24h)

table1(~sofa+sapsii|language, data = df_24h)
```

```{r language_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = language, y=sofa, col=language)) +
        theme(legend.position = "none") +
        labs(x = "Language", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = language, y=sapsii, col=language)) +
        theme(legend.position = "none") +
        labs(x = "Language", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "language"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)
```

```{r, echo=FALSE, results='asis'}
# x = "language"
# baselines = var_name_list$baselines[! var_name_list$baselines %in% c("LANGUAGE_ENGL")]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# 
# glmfit_0_rlt_list_language <- list()
# glmfit_sofa_rlt_list_language  <- list()
# glmfit_saps_rlt_list_language  <- list()
# tmle_fit_base_rlt_list_language  <- list()
# tmle_fit_base_sofa_rlt_list_language  <- list()
# tmle_fit_base_saps_rlt_list_language  <- list()
# 
# for (i in 1:length(y_vars)) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_language[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_language[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_language[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_language[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_language[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_language[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("\n haha\n\n")
# }
# save.image(here('04_24h_language.RData'))

```




## 0. TMLE estimations plots
```{r}
load(here('data', 'rdata', '04_24h_language.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_language_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_language) %>% as.data.frame()

tmle_fit_base_sofa_rlt_language_all_tsm <- tmle_fit_base_sofa_rlt_language_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_language_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_language_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_language_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```

```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))

  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) +
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)


  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_language[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_language[[i]])
  print(formatted_table)


  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_language[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_language[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_language[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_language[[i]])
  print(formatted_table)

}

```

# Partner vs Rates
```{r marital_hist}
df_baseline <- read.csv(here("data", "DataPreprocessing", "df_baseline.csv"), row.names = 1)
df_24h <- merge(df_24h, df_baseline %>% select(ICUSTAY_ID, MARITAL_STATUS), by="ICUSTAY_ID", all.x = T )

df_24h$marital_status = tolower(df_24h$MARITAL_STATUS)
df_24h$marital_status[is.na(df_24h$marital_status)] = 'Missing'
df_24h$marital_status = factor(df_24h$marital_status, levels = c("partnered", "not_partnered", "Missing"))

df_24h_all$marital_status <- df_24h$marital_status


cat("Summary of marital/partner status: \n")
table1(~marital_status, data = df_24h)

table1(~sofa+sapsii|marital_status, data = df_24h)
```

```{r marital_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = marital_status, y=sofa, col=marital_status)) +
        theme(legend.position = "none") +
        labs(x = "Partner", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = marital_status, y=sapsii, col=marital_status)) +
        theme(legend.position = "none") +
        labs(x = "Partner", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "marital_status"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)
```

```{r, echo=FALSE, results='asis'}
# x = 'marital_status'
# baselines = var_name_list$baselines[! var_name_list$baselines %in% c("MARITAL_PARTN")]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# 
# glmfit_0_rlt_list_partner <- list()
# glmfit_sofa_rlt_list_partner  <- list()
# glmfit_saps_rlt_list_partner  <- list()
# tmle_fit_base_rlt_list_partner  <- list()
# tmle_fit_base_sofa_rlt_list_partner  <- list()
# tmle_fit_base_saps_rlt_list_partner  <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_partner[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_partner[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_partner[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_partner[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_partner[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_partner[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("\n haha\n\n")
# }
# save.image(here('04_24h_partner.RData'))

```





## 0. TMLE estimations plots

```{r}
load(here('data', 'rdata', '04_24h_partner.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_partner_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_partner) %>% as.data.frame()

tmle_fit_base_sofa_rlt_partner_all_tsm <- tmle_fit_base_sofa_rlt_partner_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_partner_all_tsm$group = factor(tmle_fit_base_sofa_rlt_partner_all_tsm$group, levels = c("partnered", "not_partnered", "Missing"))

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_partner_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_partner_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_partner_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```

```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))

  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) +
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)


  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_partner[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_partner[[i]])
  print(formatted_table)


  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_partner[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_partner[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_partner[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_partner[[i]])
  print(formatted_table)

}

```

# Religion vs Rates
```{r religion_hist}
df_baseline <- read.csv(here("data", "DataPreprocessing", "df_baseline.csv"), row.names = 1)
df_24h <- merge(df_24h, df_baseline %>% select(ICUSTAY_ID, RELIGION), by="ICUSTAY_ID", all.x = T )

df_24h$religion <- tolower(df_24h$RELIGION)
df_24h$religion[is.na(df_24h$religion)] <- "Missing"

df_24h$religion[df_24h$religion == 'christian'] = 'Christian'
df_24h$religion[df_24h$religion == 'other'] = 'Other'

df_24h$religion = factor(df_24h$religion, levels = c("Christian", "Other", "Missing"))
df_24h_all$religion <- df_24h$religion


cat("Summary of religion status: \n")
table1(~religion, data = df_24h)

table1(~sofa+sapsii|religion, data = df_24h)
```

```{r religion_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = religion, y=sofa, col=religion)) +
        theme(legend.position = "none") +
        labs(x = "Religion", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = religion, y=sapsii, col=religion)) +
        theme(legend.position = "none") +
        labs(x = "Religion", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "religion"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)
```

```{r, echo=FALSE, results='asis'}
# x = "religion"
# baselines = var_name_list$baselines[! var_name_list$baselines %in% c("RELIGION_Chrt")]
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs).
# 
# glmfit_0_rlt_list_religion <- list()
# glmfit_sofa_rlt_list_religion  <- list()
# glmfit_saps_rlt_list_religion  <- list()
# tmle_fit_base_rlt_list_religion  <- list()
# tmle_fit_base_sofa_rlt_list_religion  <- list()
# tmle_fit_base_saps_rlt_list_religion  <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_religion[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_religion[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_religion[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_religion[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_religion[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_religion[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("\n haha\n\n")
# }
# save.image(here('04_24h_religion.RData'))

```



## 0. TMLE estimations plots
```{r}
load(here("data", "rdata", '04_24h_religion.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_religion_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_religion) %>% as.data.frame()

tmle_fit_base_sofa_rlt_religion_all_tsm <- tmle_fit_base_sofa_rlt_religion_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_religion_all_tsm$group = factor(tmle_fit_base_sofa_rlt_religion_all_tsm$group, levels = c("Christian", "Other", "Missing"))

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_religion_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_religion_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_religion_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```

```{r, fig.width=6, fig.height=3, results='asis'}
for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))

  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))
  
  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) +
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)


  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_religion[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_religion[[i]])
  print(formatted_table)


  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_religion[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_religion[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_religion[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_religion[[i]])
  print(formatted_table)

}

```



# Care units
```{r careunit_hist}
ICUSTAYS <- read.csv(here("data", "raw_data", "mimic-iii-clinical-database-1.4", "ICUSTAYS.csv"))
df_24h <- merge(df_24h, ICUSTAYS %>% select(ICUSTAY_ID, FIRST_CAREUNIT), by="ICUSTAY_ID", all.x = T )

df_24h$FIRST_CAREUNIT <- ifelse(is.na(df_24h$FIRST_CAREUNIT), "Missing", df_24h$FIRST_CAREUNIT)
df_24h_all$FIRST_CAREUNIT <- df_24h$FIRST_CAREUNIT

cat("Summary of first care unit: \n")
table1(~FIRST_CAREUNIT, data = df_24h)

table1(~sofa+sapsii+less_24h|FIRST_CAREUNIT, data = df_24h)


table1(as.formula(paste0("~", paste(c("age_cat", "ethnicity", "gender", "insurance", "language", "marital_status", "religion"), collapse=" + "), "|FIRST_CAREUNIT")) , data = df_24h_all)
table1(as.formula(paste0("~", paste(var_name_list$hours_missing_vitals, collapse=" + "), "|FIRST_CAREUNIT")) , data = df_24h)
table1(as.formula(paste0("~", paste(var_name_list$num_vitals, collapse=" + "), "|FIRST_CAREUNIT")) , data = df_24h)
table1(as.formula(paste0("~", paste(var_name_list$num_labs, collapse=" + "), "|FIRST_CAREUNIT")) , data = df_24h)
```

```{r, fig.height=3, fig.width=16}
x = "FIRST_CAREUNIT"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)


g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)
```


```{r careunit_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = FIRST_CAREUNIT, y=sofa, col=FIRST_CAREUNIT)) +
        theme(legend.position = "none") +
        labs(x = "First Care Unit", y = "score", title = "SOFA")
p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = FIRST_CAREUNIT, y=sapsii, col=FIRST_CAREUNIT)) +
        theme(legend.position = "none") +
        labs(x = "First Care Unit", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, echo=FALSE, results='asis'}
# x = "FIRST_CAREUNIT"
# baselines = var_name_list$baselines
# y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
# 
# glmfit_0_rlt_list_careunit <- list()
# glmfit_sofa_rlt_list_careunit  <- list()
# glmfit_saps_rlt_list_careunit  <- list()
# tmle_fit_base_rlt_list_careunit  <- list()
# tmle_fit_base_sofa_rlt_list_careunit  <- list()
# tmle_fit_base_saps_rlt_list_careunit  <- list()
# 
# for (i in 1:length(y_vars)) {
# # for (i in 16:18) {
#   y = y_vars[i]
#   cat(paste0("\n## ", i, ". ", y, "\n\n"))
# 
#   source(' 04_rates_24h_template.R')
# 
#   glmfit_0_rlt_list_careunit[[i]] <- glmfit_0_rlt
#   glmfit_sofa_rlt_list_careunit[[i]] <- glmfit_sofa_rlt
#   glmfit_saps_rlt_list_careunit[[i]] <- glmfit_saps_rlt
#   tmle_fit_base_rlt_list_careunit[[i]] <- tmle_fit_base_rlt
#   tmle_fit_base_sofa_rlt_list_careunit[[i]] <- tmle_fit_base_sofa_rlt
#   tmle_fit_base_saps_rlt_list_careunit[[i]] <- tmle_fit_base_saps_rlt
# 
#   print("\n haha\n\n")
# }
# save.image(here('04_24h_careunit.RData'))

```




## 0. TMLE estimations plots
```{r}
load(here('data', 'rdata', '04_24h_careunit.RData'))
source(here("scripts", "04_rates_24h_functions.R"))

plotFolder <- here("results","images", "04_24h")
```


```{r fig.width=16, fig.height=3}
tmle_fit_base_sofa_rlt_careunit_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_careunit) %>% as.data.frame()

tmle_fit_base_sofa_rlt_careunit_all_tsm <- tmle_fit_base_sofa_rlt_careunit_all %>%
  filter(type == 'TSM') %>%
  mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param),
         Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",
                         ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital",
                                ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))),
         Y_name = sub("^(n_|h_m_)", "", Y))

tmle_fit_base_sofa_rlt_careunit_all_tsm$group = factor(tmle_fit_base_sofa_rlt_careunit_all_tsm$group)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_careunit_all_tsm, 'h_m_vital') +
  labs(title = "Vital Sign Missing Hours")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200)


g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_careunit_all_tsm, 'n_vital') +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200)

g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_careunit_all_tsm, 'n_lab') +
  labs(title = "Laboratory Testing Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200)
```

```{r, fig.width=6, fig.height=3, results='asis'}

df_24h$FIRST_CAREUNIT <- factor(df_24h$FIRST_CAREUNIT)

for (i in 1:length(y_vars)) {

  y = y_vars[i]
  cat(paste0("\n\n## ", i, ". ", y, "\n\n"))

  cat("Summary of", y,  "w.r.t.", x, "groups: \n")
  table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h)
  cat(knitr::asis_output(table_result))

  p <- ggplot(data = df_24h) +
    geom_density(aes_string(x=y, fill=x), binwidth=1) +
    facet_wrap(df_24h[,x])
  ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200)
  plot(p)


  cat('\n\n### 1. No confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_0_rlt_list_careunit[[i]])
  print(formatted_table)

  cat('\n\n### 2. SOFA score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_sofa_rlt_list_careunit[[i]])
  print(formatted_table)


  cat('\n\n### 3. SAPS-II score confounder:\n')
  cat("Summary of linear regression w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(glmfit_saps_rlt_list_careunit[[i]])
  print(formatted_table)

  cat('\n\n### 4. Baseline confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_rlt_list_careunit[[i]])
  print(formatted_table)

  cat('\n\n### 5. Baselineb & SOFA score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_careunit[[i]])
  print(formatted_table)

  cat('\n\n### 6. Baseline & SAPSII score confounders:\n')
  cat("Summary of TMLE fit w.r.t.", x, "groups: \n")
  formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_careunit[[i]])
  print(formatted_table)

}

```


```{r}
knitr::knit_exit()
```

# Age & Ethnicity vs rates
## Age & Ethnicity distribution
```{r}
df_24h$age_cat_2 = ifelse(df_24h$age_cat %in% c("18-30", "31-45"), "18-45", ">45")

df_24h$ethnicity[is.na(df_24h$ethnicity)] = "Missing"

df_24h$age_eth = paste0(df_24h$age_cat_2, "_", df_24h$ethnicity)

df_24h$age_eth = factor(df_24h$age_eth , levels = c("18-45_WHITE", "18-45_ASIAN", "18-45_BLACK", "18-45_HISPANIC", "18-45_Other", "18-45_Missing",
                                                    ">45_WHITE", ">45_ASIAN", ">45_BLACK", ">45_HISPANIC", ">45_Other", ">45_Missing"))
df_24h_all$age_eth <- df_24h$age_eth

cat("Summary of age + ethnicity distribution: \n")
table1(~age_eth, data = df_24h)

table1(~sofa+sapsii|age_eth, data = df_24h)
```


```{r age_ethnicity_severity_Scores, fig.height=3, fig.width=8}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = age_eth, y=sofa, col=age_eth)) +
        theme(legend.position = "none",
              axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0)) +
        labs(x = "Age & Ethnicity group", y = "score", title = "SOFA")

p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = age_eth, y=sapsii, col=age_eth)) +
        theme(legend.position = "none",
              axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0)) +
        labs(x = "Age & Ethnicity group", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "age_eth"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)

```


```{r}
x = "age_eth"
baselines = var_name_list$baselines[! var_name_list$baselines %in% c("AGE", "ETHNICITY_W", "ETHNICITY_B", "ETHNICITY_A", "ETHNICITY_H", "ETHNICITY_O")]
y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
idx <- 1:length(y_vars)

glmfit_0_rlt_list_age_eth <- list()
glmfit_sofa_rlt_list_age_eth  <- list()
glmfit_saps_rlt_list_age_eth  <- list()
tmle_fit_base_rlt_list_age_eth  <- list()
tmle_fit_base_sofa_rlt_list_age_eth  <- list()
tmle_fit_base_saps_rlt_list_age_eth  <- list()

for (i in 1:length(y_vars)) {
# for (i in 16:18) {
  y = y_vars[i]
  cat(paste0("\n## ", i, ". ", y, "\n\n"))

  source(' 04_rates_24h_template.R')

  glmfit_0_rlt_list_age_eth[[i]] <- glmfit_0_rlt
  glmfit_sofa_rlt_list_age_eth[[i]] <- glmfit_sofa_rlt
  glmfit_saps_rlt_list_age_eth[[i]] <- glmfit_saps_rlt
  tmle_fit_base_rlt_list_age_eth[[i]] <- tmle_fit_base_rlt
  tmle_fit_base_sofa_rlt_list_age_eth[[i]] <- tmle_fit_base_sofa_rlt
  tmle_fit_base_saps_rlt_list_age_eth[[i]] <- tmle_fit_base_saps_rlt

  print("haha\n\n")
}
save.image(here('04_24h_age_eth.RData'))

```


<!-- ## 0. TMLE estimations plots  -->

<!-- ```{r} -->
<!-- load(here('04_24h_age_eth.RData')) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- make_box_plots <- function(df_box, y_vars, x){ -->
<!--   meltData <- melt(df_box %>% dplyr::select(y_vars, x), id.vars = x)  -->
<!--   names(meltData)[1] = "x" -->
<!--   if(x == 'age_cat'){ -->
<!--     meltData$x <- factor(meltData$x, levels = c('18-30', '31-45', '46-65', '>65'))  -->
<!--   } else { -->
<!--     meltData$x <- as.factor(meltData$x) -->
<!--   } -->
<!--   meltData$item = sub("^(n_|h_m_)", "", meltData$variable) -->
<!--   pi <- ggplot(meltData, aes(x = x, y = value, color = x)) + -->
<!--     geom_boxplot(position = position_dodge(width = 1)) +   -->
<!--     facet_wrap(~ item, scales = 'free', nrow = 1) + -->
<!--     theme_bw() + -->
<!--     theme(legend.position = 'none', -->
<!--         axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0), -->
<!--         axis.title.x = element_blank(), -->
<!--         plot.title = element_text(hjust = 0.5), -->
<!--         plot.subtitle = element_text(hjust = 0.5)) + -->
<!--     labs(color="x", x = "", y = "count") -->
<!--   return(pi) -->
<!-- } -->



<!-- make_box_plots_tmle_rlts <- function(df_box, y_type){ -->
<!--   pi <- ggplot(df_box %>% filter(Y_type == y_type),  -->
<!--        aes(x = group)) + -->
<!--   geom_errorbar(aes(ymin=lower, ymax=upper, color=group), width=0.7) + -->
<!--   geom_point(aes(y=tmle_est, color = group))+ -->
<!--   facet_wrap(~ Y_name, scales = 'free', nrow = 1) + -->
<!--   labs(y="TMLE estimated rates",  -->
<!--        subtitle = 'TMLE estimation with other baseline variables and SOFA score as confounders') + -->
<!--   theme_bw() + -->
<!--   theme(legend.position = 'none', -->
<!--         axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0), -->
<!--         axis.title.x = element_blank(), -->
<!--         plot.title = element_text(hjust = 0.5), -->
<!--         plot.subtitle = element_text(hjust = 0.5))  -->

<!--   return(pi) -->
<!-- } -->



<!-- ``` -->


<!-- ```{r fig.width=16, fig.height=3} -->
<!-- tmle_fit_base_sofa_rlt_age_eth_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_age_eth) %>% as.data.frame() -->

<!-- tmle_fit_base_sofa_rlt_age_eth_all_tsm <- tmle_fit_base_sofa_rlt_age_eth_all %>% -->
<!--   filter(type == 'TSM') %>% -->
<!--   mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param), -->
<!--          Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",  -->
<!--                          ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital", -->
<!--                                 ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))), -->
<!--          Y_name = sub("^(n_|h_m_)", "", Y)) -->

<!-- tmle_fit_base_sofa_rlt_age_eth_all_tsm$group = factor(tmle_fit_base_sofa_rlt_age_eth_all_tsm$group, levels = c("18-45_WHITE", "18-45_ASIAN", "18-45_BLACK", "18-45_HISPANIC", "18-45_Other", "18-45_Missing", -->
<!--                                                     ">45_WHITE", ">45_ASIAN", ">45_BLACK", ">45_HISPANIC", ">45_Other", ">45_Missing")) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_eth_all_tsm, 'h_m_vital') + -->
<!--   labs(title = "Vital Sign Missing Hours")  -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_eth_all_tsm, 'n_vital') + -->
<!--   labs(title = "Vital Sign Measurement Frequencies")  -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_eth_all_tsm, 'n_lab') + -->
<!--   labs(title = "Laboratory Testing Frequencies")  -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200) -->
<!-- ``` -->

<!-- ```{r, fig.width=6, fig.height=3, results='asis'} -->
<!-- for (i in 1:length(y_vars)) { -->

<!--   y = y_vars[i] -->
<!--   cat(paste0("\n\n## ", i, ". ", y, "\n\n")) -->

<!--   cat("Summary of", y,  "w.r.t.", x, "groups: \n")   -->
<!--   table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h) -->
<!--   cat(knitr::asis_output(table_result)) -->

<!--   p <- ggplot(data = df_24h) + -->
<!--     geom_density(aes_string(x=y, fill=x), binwidth=1) +  -->
<!--     facet_wrap(df_24h[,x]) -->
<!--   ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200) -->
<!--   plot(p) -->


<!--   cat('\n\n### 1. No confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_0_rlt_list_age_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 2. SOFA score confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_sofa_rlt_list_age_eth[[i]]) -->
<!--   print(formatted_table) -->


<!--   cat('\n\n### 3. SAPS-II score confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_saps_rlt_list_age_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 4. Baseline confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_rlt_list_age_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 5. Baselineb & SOFA score confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_age_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 6. Baseline & SAPSII score confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_age_eth[[i]]) -->
<!--   print(formatted_table) -->
<!-- } -->

<!-- ``` -->



# Age & Gender vs rates
## Age & Gender distribution
```{r}
df_24h$age_cat_2 = ifelse(df_24h$age_cat %in% c("18-30", "31-45"), "18-45", ">45")
df_24h$age_gender = paste0(df_24h$age_cat_2, "_", df_24h$gender)

df_24h$age_gender = factor(df_24h$age_gender , levels = c("18-45_Female", "18-45_Male", 
                                                    ">45_Female", ">45_Male"))
df_24h_all$age_gender <- df_24h$age_gender

cat("Summary of age + gender distribution: \n")
table1(~age_gender, data = df_24h)

table1(~sofa+sapsii|age_gender, data = df_24h)
```


```{r age_gender_severity_Scores, fig.height=3, fig.width=7}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = age_gender, y=sofa, col=age_gender)) +
        theme(legend.position = "none",
              axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0)) +
        labs(x = "Age & Gender group", y = "score", title = "SOFA")

p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = age_gender, y=sapsii, col=age_gender)) +
        theme(legend.position = "none",
              axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0)) +
        labs(x = "Age & Gender group", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "age_gender"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)

```


```{r}
x = "age_gender"
baselines = var_name_list$baselines[! var_name_list$baselines %in% c("AGE", "GENDER_F")]
y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
idx <- 1:length(y_vars)

glmfit_0_rlt_list_age_gender <- list()
glmfit_sofa_rlt_list_age_gender  <- list()
glmfit_saps_rlt_list_age_gender  <- list()
tmle_fit_base_rlt_list_age_gender  <- list()
tmle_fit_base_sofa_rlt_list_age_gender  <- list()
tmle_fit_base_saps_rlt_list_age_gender  <- list()

for (i in 1:length(y_vars)) {
# for (i in 16:18) {
  y = y_vars[i]
  cat(paste0("\n## ", i, ". ", y, "\n\n"))

  source(' 04_rates_24h_template.R')

  glmfit_0_rlt_list_age_gender[[i]] <- glmfit_0_rlt
  glmfit_sofa_rlt_list_age_gender[[i]] <- glmfit_sofa_rlt
  glmfit_saps_rlt_list_age_gender[[i]] <- glmfit_saps_rlt
  tmle_fit_base_rlt_list_age_gender[[i]] <- tmle_fit_base_rlt
  tmle_fit_base_sofa_rlt_list_age_gender[[i]] <- tmle_fit_base_sofa_rlt
  tmle_fit_base_saps_rlt_list_age_gender[[i]] <- tmle_fit_base_saps_rlt

  print("haha\n\n")
}
save.image(here('04_24h_age_gender.RData'))

```


<!-- ## 0. TMLE estimations plots  -->

<!-- ```{r} -->
<!-- load(here('04_24h_age_gender.RData')) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- make_box_plots <- function(df_box, y_vars, x){ -->
<!--   meltData <- melt(df_box %>% dplyr::select(y_vars, x), id.vars = x)  -->
<!--   names(meltData)[1] = "x" -->
<!--   if(x == 'age_cat'){ -->
<!--     meltData$x <- factor(meltData$x, levels = c('18-30', '31-45', '46-65', '>65'))  -->
<!--   } else { -->
<!--     meltData$x <- as.factor(meltData$x) -->
<!--   } -->
<!--   meltData$item = sub("^(n_|h_m_)", "", meltData$variable) -->
<!--   pi <- ggplot(meltData, aes(x = x, y = value, color = x)) + -->
<!--     geom_boxplot(position = position_dodge(width = 1)) +   -->
<!--     facet_wrap(~ item, scales = 'free', nrow = 1) + -->
<!--     theme_bw() + -->
<!--     theme(legend.position = 'none', -->
<!--         axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0), -->
<!--         axis.title.x = element_blank(), -->
<!--         plot.title = element_text(hjust = 0.5), -->
<!--         plot.subtitle = element_text(hjust = 0.5)) + -->
<!--     labs(color="x", x = "", y = "count") -->
<!--   return(pi) -->
<!-- } -->



<!-- make_box_plots_tmle_rlts <- function(df_box, y_type){ -->
<!--   pi <- ggplot(df_box %>% filter(Y_type == y_type),  -->
<!--        aes(x = group)) + -->
<!--   geom_errorbar(aes(ymin=lower, ymax=upper, color=group), width=0.7) + -->
<!--   geom_point(aes(y=tmle_est, color = group))+ -->
<!--   facet_wrap(~ Y_name, scales = 'free', nrow = 1) + -->
<!--   labs(y="TMLE estimated rates",  -->
<!--        subtitle = 'TMLE estimation with other baseline variables and SOFA score as confounders') + -->
<!--   theme_bw() + -->
<!--   theme(legend.position = 'none', -->
<!--         axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0), -->
<!--         axis.title.x = element_blank(), -->
<!--         plot.title = element_text(hjust = 0.5), -->
<!--         plot.subtitle = element_text(hjust = 0.5))  -->

<!--   return(pi) -->
<!-- } -->



<!-- ``` -->


<!-- ```{r fig.width=16, fig.height=3} -->
<!-- tmle_fit_base_sofa_rlt_age_gender_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_age_gender) %>% as.data.frame() -->

<!-- tmle_fit_base_sofa_rlt_age_gender_all_tsm <- tmle_fit_base_sofa_rlt_age_gender_all %>% -->
<!--   filter(type == 'TSM') %>% -->
<!--   mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param), -->
<!--          Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital",  -->
<!--                          ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital", -->
<!--                                 ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))), -->
<!--          Y_name = sub("^(n_|h_m_)", "", Y)) -->

<!-- tmle_fit_base_sofa_rlt_age_gender_all_tsm$group = factor(tmle_fit_base_sofa_rlt_age_gender_all_tsm$group, levels = c("18-45_Female", "18-45_Male",  -->
<!--                                                     ">45_Female", ">45_Male")) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_gender_all_tsm, 'h_m_vital') + -->
<!--   labs(title = "Vital Sign Missing Hours")  -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_gender_all_tsm, 'n_vital') + -->
<!--   labs(title = "Vital Sign Measurement Frequencies")  -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_age_gender_all_tsm, 'n_lab') + -->
<!--   labs(title = "Laboratory Testing Frequencies")  -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200) -->
<!-- ``` -->

<!-- ```{r, fig.width=6, fig.height=3, results='asis'} -->
<!-- for (i in 1:length(y_vars)) { -->

<!--   y = y_vars[i] -->
<!--   cat(paste0("\n\n## ", i, ". ", y, "\n\n")) -->

<!--   cat("Summary of", y,  "w.r.t.", x, "groups: \n")   -->
<!--   table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h) -->
<!--   cat(knitr::asis_output(table_result)) -->

<!--   p <- ggplot(data = df_24h) + -->
<!--     geom_density(aes_string(x=y, fill=x), binwidth=1) +  -->
<!--     facet_wrap(df_24h[,x]) -->
<!--   ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200) -->
<!--   plot(p) -->


<!--   cat('\n\n### 1. No confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_0_rlt_list_age_gender[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 2. SOFA score confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_sofa_rlt_list_age_gender[[i]]) -->
<!--   print(formatted_table) -->


<!--   cat('\n\n### 3. SAPS-II score confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_saps_rlt_list_age_gender[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 4. Baseline confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_rlt_list_age_gender[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 5. Baselineb & SOFA score confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_age_gender[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 6. Baseline & SAPSII score confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_age_gender[[i]]) -->
<!--   print(formatted_table) -->
<!-- } -->

<!-- ``` -->





# Gender & Ethnicity vs rates
## Gender & Ethnicity distribution
```{r}
df_24h$ethnicity[is.na(df_24h$ethnicity)] = "Missing"
df_24h$gender_eth = paste0(df_24h$gender, "_", df_24h$ethnicity)

df_24h$gender_eth = factor(df_24h$gender_eth , levels = c("Female_WHITE", "Female_ASIAN", "Female_BLACK", "Female_HISPANIC", "Female_Other", "Female_Missing",
                                                          "Male_WHITE", "Male_ASIAN", "Male_BLACK", "Male_HISPANIC", "Male_Other", "Male_Missing"))
df_24h_all$gender_eth <- df_24h$gender_eth

cat("Summary of gender + ethnicity distribution: \n")
table1(~gender_eth, data = df_24h)

table1(~sofa+sapsii|gender_eth, data = df_24h)
```


```{r gender_ethnicity_severity_Scores, fig.height=3, fig.width=7}
p1 <- ggplot(df_24h) +
        geom_boxplot(aes(x = gender_eth, y=sofa, col=gender_eth)) +
        theme(legend.position = "none",
              axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0)) +
        labs(x = "Gender & Ethnicity group", y = "score", title = "SOFA")

p2 <- ggplot(df_24h) +
        geom_boxplot(aes(x = gender_eth, y=sapsii, col=gender_eth)) +
        theme(legend.position = "none",
              axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0)) +
        labs(x = "Gender & Ethnicity group", y = "score", title = "SAPS-II")
grid.arrange(p1, p2, nrow = 1)
```


```{r, fig.height=3, fig.width=16}
x = "gender_eth"

g <- make_box_plots(df_24h, var_name_list$hours_missing_vitals, x) +
  labs(title = "Vital Sign Missing Hours") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_h_m_.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_vitals, x) +
  labs(title = "Vital Sign Measurement Frequencies")
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_vital_n.png")), g, width = 11, height = 3, dpi = 1200)

g <- make_box_plots(df_24h, var_name_list$num_labs, x) +
  labs(title = "Lab Testing Frequencies") 
g
ggsave(file=here(plotFolder, paste0(x, "_box_plot_lab_n.png")), g, width = 11, height = 3, dpi = 1200)

```


```{r}
x = "gender_eth"
baselines = var_name_list$baselines[! var_name_list$baselines %in% c("AGE", "GENDER_F")]
y_vars <- c(var_name_list$num_vitals, var_name_list$hours_missing_vitals, var_name_list$num_labs)
idx <- 1:length(y_vars)

glmfit_0_rlt_list_gender_eth <- list()
glmfit_sofa_rlt_list_gender_eth  <- list()
glmfit_saps_rlt_list_gender_eth  <- list()
tmle_fit_base_rlt_list_gender_eth  <- list()
tmle_fit_base_sofa_rlt_list_gender_eth  <- list()
tmle_fit_base_saps_rlt_list_gender_eth  <- list()

for (i in 1:length(y_vars)) {
# for (i in 16:18) {
  y = y_vars[i]
  cat(paste0("\n## ", i, ". ", y, "\n\n"))

  source(' 04_rates_24h_template.R')

  glmfit_0_rlt_list_gender_eth[[i]] <- glmfit_0_rlt
  glmfit_sofa_rlt_list_gender_eth[[i]] <- glmfit_sofa_rlt
  glmfit_saps_rlt_list_gender_eth[[i]] <- glmfit_saps_rlt
  tmle_fit_base_rlt_list_gender_eth[[i]] <- tmle_fit_base_rlt
  tmle_fit_base_sofa_rlt_list_gender_eth[[i]] <- tmle_fit_base_sofa_rlt
  tmle_fit_base_saps_rlt_list_gender_eth[[i]] <- tmle_fit_base_saps_rlt

  print("haha\n\n")
}
save.image(here('04_24h_gender_eth.RData'))

```


<!-- ## 0. TMLE estimations plots -->

<!-- ```{r} -->
<!-- load(here('04_24h_gender_eth.RData')) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- make_box_plots <- function(df_box, y_vars, x){ -->
<!--   meltData <- melt(df_box %>% dplyr::select(y_vars, x), id.vars = x) -->
<!--   names(meltData)[1] = "x" -->
<!--   if(x == 'age_cat'){ -->
<!--     meltData$x <- factor(meltData$x, levels = c('18-30', '31-45', '46-65', '>65')) -->
<!--   } else { -->
<!--     meltData$x <- as.factor(meltData$x) -->
<!--   } -->
<!--   meltData$item = sub("^(n_|h_m_)", "", meltData$variable) -->
<!--   pi <- ggplot(meltData, aes(x = x, y = value, color = x)) + -->
<!--     geom_boxplot(position = position_dodge(width = 1)) + -->
<!--     facet_wrap(~ item, scales = 'free', nrow = 1) + -->
<!--     theme_bw() + -->
<!--     theme(legend.position = 'none', -->
<!--         axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0), -->
<!--         axis.title.x = element_blank(), -->
<!--         plot.title = element_text(hjust = 0.5), -->
<!--         plot.subtitle = element_text(hjust = 0.5)) + -->
<!--     labs(color="x", x = "", y = "count") -->
<!--   return(pi) -->
<!-- } -->



<!-- make_box_plots_tmle_rlts <- function(df_box, y_type){ -->
<!--   pi <- ggplot(df_box %>% filter(Y_type == y_type), -->
<!--        aes(x = group)) + -->
<!--   geom_errorbar(aes(ymin=lower, ymax=upper, color=group), width=0.7) + -->
<!--   geom_point(aes(y=tmle_est, color = group))+ -->
<!--   facet_wrap(~ Y_name, scales = 'free', nrow = 1) + -->
<!--   labs(y="TMLE estimated rates", -->
<!--        subtitle = 'TMLE estimation with other baseline variables and SOFA score as confounders') + -->
<!--   theme_bw() + -->
<!--   theme(legend.position = 'none', -->
<!--         axis.text.x = element_text(angle = -75, vjust = 0.5, hjust=0), -->
<!--         axis.title.x = element_blank(), -->
<!--         plot.title = element_text(hjust = 0.5), -->
<!--         plot.subtitle = element_text(hjust = 0.5)) -->

<!--   return(pi) -->
<!-- } -->



<!-- ``` -->


<!-- ```{r fig.width=16, fig.height=3} -->
<!-- tmle_fit_base_sofa_rlt_gender_eth_all <- do.call("rbind", tmle_fit_base_sofa_rlt_list_gender_eth) %>% as.data.frame() -->

<!-- tmle_fit_base_sofa_rlt_gender_eth_all_tsm <- tmle_fit_base_sofa_rlt_gender_eth_all %>% -->
<!--   filter(type == 'TSM') %>% -->
<!--   mutate(group = gsub(".*A=(.*)\\}.*", "\\1", param), -->
<!--          Y_type = ifelse(Y %in% var_name_list$num_vitals, "n_vital", -->
<!--                          ifelse(Y %in% var_name_list$hours_missing_vitals, "h_m_vital", -->
<!--                                 ifelse(Y %in% var_name_list$num_labs, "n_lab", 'other'))), -->
<!--          Y_name = sub("^(n_|h_m_)", "", Y)) -->

<!-- tmle_fit_base_sofa_rlt_gender_eth_all_tsm$group = factor(tmle_fit_base_sofa_rlt_gender_eth_all_tsm$group, levels = c("18-45_Female", "18-45_Male", -->
<!--                                                     ">45_Female", ">45_Male")) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_gender_eth_all_tsm, 'h_m_vital') + -->
<!--   labs(title = "Vital Sign Missing Hours") -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_h_m.png")), g, width = 15, height = 3, dpi = 1200) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_gender_eth_all_tsm, 'n_vital') + -->
<!--   labs(title = "Vital Sign Measurement Frequencies") -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_vital_n.png")), g, width = 15, height = 3, dpi = 1200) -->

<!-- g <- make_box_plots_tmle_rlts(tmle_fit_base_sofa_rlt_gender_eth_all_tsm, 'n_lab') + -->
<!--   labs(title = "Laboratory Testing Frequencies") -->
<!-- g -->
<!-- ggsave(file=here(plotFolder, paste0(x, "_tmle_rlts_box_plot_lab_n.png")), g, width = 15, height = 3, dpi = 1200) -->
<!-- ``` -->

<!-- ```{r, fig.width=6, fig.height=3, results='asis'} -->
<!-- for (i in 1:length(y_vars)) { -->

<!--   y = y_vars[i] -->
<!--   cat(paste0("\n\n## ", i, ". ", y, "\n\n")) -->

<!--   cat("Summary of", y,  "w.r.t.", x, "groups: \n") -->
<!--   table_result <- table1(as.formula(paste0("~ ", y, "| ", x)), data = df_24h) -->
<!--   cat(knitr::asis_output(table_result)) -->

<!--   p <- ggplot(data = df_24h) + -->
<!--     geom_density(aes_string(x=y, fill=x), binwidth=1) + -->
<!--     facet_wrap(df_24h[,x]) -->
<!--   ggsave(file=here(plotFolder, paste0(x, "_", i, "_", y, "_EDA.png")), p, width = 6, height = 3, dpi = 1200) -->
<!--   plot(p) -->


<!--   cat('\n\n### 1. No confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_0_rlt_list_gender_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 2. SOFA score confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_sofa_rlt_list_gender_eth[[i]]) -->
<!--   print(formatted_table) -->


<!--   cat('\n\n### 3. SAPS-II score confounder:\n') -->
<!--   cat("Summary of linear regression w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(glmfit_saps_rlt_list_gender_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 4. Baseline confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_rlt_list_gender_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 5. Baselineb & SOFA score confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_sofa_rlt_list_gender_eth[[i]]) -->
<!--   print(formatted_table) -->

<!--   cat('\n\n### 6. Baseline & SAPSII score confounders:\n') -->
<!--   cat("Summary of TMLE fit w.r.t.", x, "groups: \n") -->
<!--   formatted_table <- knitr::kable(tmle_fit_base_saps_rlt_list_gender_eth[[i]]) -->
<!--   print(formatted_table) -->
<!-- } -->

<!-- ``` -->
