==> scripts/05_VIA.Rmd <==
---
title: "VIA"
author: "Seraphina Shi"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r load_lib, include = FALSE, warning=FALSE, message=FALSE, echo=FALSE} 
# setwd("/home/seraphinashi/MissingDataInTheICU")
# install.packages("polspline")
library(dplyr)
library(data.table)
library(SuperLearner)
library(origami)
library(sl3)
library(ggplot2)
library(forcats)
library(here)
library(future)
library(ROCR) 
library(pROC)
library(PRROC)
library(Metrics)
```


```{r setup, include = FALSE}
plotFolder <- here("results","images","05_VIA")
if(!file.exists(plotFolder)) dir.create(plotFolder,recursive=TRUE)

rdataFolder <- here("data", "rdata")

knitr::opts_chunk$set(
  cache=FALSE, autodep=FALSE, warning=FALSE, message=FALSE, echo=FALSE,
  results = 'markup', dev='png', dpi=150, fig.align = "center", fig.path=plotFolder,
  cache.path=".cache/",
  duplicate.label="allow"
)

set.seed(123)

```


# Get data
```{r}
load(here(rdataFolder, "02_2_varFusion.RData"))

PreprocessDataFolder <- here("data", "DataPreprocessing")

df <- read.csv(here(PreprocessDataFolder,"02_3_data_complt_blocks.csv"), row.names = 1, header = T)  
df_imp <- read.csv(here(PreprocessDataFolder,"02_3_data_imputed_complt_blocks.csv"), row.names = 1, header = T)
df_r_imp <- read.csv(here(PreprocessDataFolder,"02_3_data_random_imputed_complt_blocks.csv"), row.names = 1, header = T)
```

```{r}
# discharged within the first 24 hours
df_tmp <- df[df$half_day <= 2, ]
discharge_in24_id <- df_tmp$ICUSTAY_ID[df_tmp$Y_t == 2]

df <- df[! df$ICUSTAY_ID %in% discharge_in24_id, ]
df_imp <- df_imp[! df_imp$ICUSTAY_ID %in% discharge_in24_id, ]
df_r_imp <- df_r_imp[! df_r_imp$ICUSTAY_ID %in% discharge_in24_id, ]
```



```{r}
cat("Missing value percentage for each column in the data: \n")
cat("  Baselines: \n")
print(round((colMeans(is.na(df %>% select(all_of(var_name_list$baselines)))))*100,4))

cat("  Vital signs: \n")
print(round((colMeans(is.na(df %>% select(var_name_list$vitals))))*100,4))
cat("  Lab tests: \n")
print(round((colMeans(is.na(df %>% select(var_name_list$labs))))*100,4))
```


# Train predictive models with current 12 hour time block data
```{r}
get_df_prev <- function(df){
    df_prev <- df %>% 
      select(c(var_name_list$ID, 
               var_name_list$vitals, var_name_list$hours_missing_vitals, var_name_list$num_vitals,
               var_name_list$labs, var_name_list$num_labs)) 
    names(df_prev)[-c(1:2)] <- paste0(names(df_prev)[-c(1:2)], "_prev")
    df_prev$half_day_prev <- df_prev$half_day 
    df_prev$half_day <- df_prev$half_day + 1
    
    df_prev  <- df_prev[df_prev$half_day <= 10, ]
    df <- merge(df, df_prev, by = c("ICUSTAY_ID", "half_day"), all.x = T)
    df <- df[df$half_day != 1, ]
    
    return(df)
}


set.seed(123)
smaller_id_set <- c(sample(unique(df$ICUSTAY_ID[df$Y_t_plus_1 == 0]), 3000), df$ICUSTAY_ID[df$Y_t_plus_1 == 1])
smaller_df_idx <- which(df$ICUSTAY_ID %in% smaller_id_set)

get_df_sl <- function(sets, smaller_df = T, prev_block = F, inputation = "median"){
  if(smaller_df){
    dat <- df[smaller_df_idx, ]
    dat_imp <- df_imp[smaller_df_idx,]
    dat_r_imp <- df_r_imp[smaller_df_idx, ]
  } else {
    dat <- df
    dat_imp <- df_imp
    dat_r_imp <- df_r_imp
  }

  
  if(prev_block){
    df_prev <- get_df_prev(dat)
    dat_imp_prev <- get_df_prev(dat_imp)
    dat_r_imp <- get_df_prev(df_r_imp)
  }
  
  vars <- c()
  
  if("VS" %in% sets)  vars <- c(vars, var_name_list$vitals)
  if("h_m_VS" %in% sets)  vars <- c(vars, var_name_list$hours_missing_vitals)
  if("n_VS" %in% sets)  vars <- c(vars, var_name_list$num_vitals)
  if("LT" %in% sets)  vars <- c(vars, var_name_list$labs)
  if("n_LT" %in% sets)  vars <- c(vars, var_name_list$num_labs)
  
  if(prev_block){
    if(length(vars) > 0){
      vars <- c(vars, paste0(vars, "_prev"))
    }
  }
  
  if("W" %in% sets)  vars <- c(vars, var_name_list$baselines)
  if("n_W" %in% sets)  vars <- c(vars, var_name_list$num_baselines)
  vars <- c(var_name_list$ID,
            var_name_list$outcome,
            vars)
  
  df_sl3 <- dat %>% dplyr::select(vars)
  
  if(inputation == "CCA"){
    df_sl3 <- df_sl3[complete.cases(df_sl3), ] 
  } else if(inputation == "random_unif"){
    df_sl3 <- dat_r_imp %>% dplyr::select(vars)
  } else {
    df_sl3 <- dat_imp %>% dplyr::select(vars)
  }
  return(df_sl3)
}

run_sl <- function(df_in, rf_only = T, check_VIA = F){
    
  df_here <- df_in %>% select(-half_day)
  
  ## define sl3 task
  
  # define the folds with clusters and stratification
  set.seed(123)
  folds <- make_folds(nrow(df_here), cluster_ids = df_here$ICUSTAY_ID)
  
  # define the prediction task
  sl3_task <- make_sl3_Task(
    data = df_here, 
    covariates = colnames(df_here)[-which(names(df_here) %in% c(var_name_list$outcome,
                                                              var_name_list$ID))],
    outcome = var_name_list$outcome,
    outcome_type = "binomial",
    id = "ICUSTAY_ID",
    folds = folds
  )
  #====================
  
  ######################### build library for discrete SL ########################
  
  ######## learners without screeners
  if (rf_only){
    lrn_mean <- Lrnr_mean$new()
    lrn_ranger <- Lrnr_ranger$new()

    candidate_lrnrs <- c(lrn_mean, lrn_ranger)
    names(candidate_lrnrs) <- c("Mean", "Ranger")
    candidate_lrnrs_stack <- Stack$new(candidate_lrnrs) 
    
  } else {
    # ,  , , , , , , , and the Ensemble SL itself

    lrn_glm <- Lrnr_glm$new() # Generalized Linear Model (GLM)
    lrn_earth <- Lrnr_earth$new() # Generalized Additive Model (GAM)
    lrn_ranger <- Lrnr_ranger$new() # Random Forests, 
    lrn_bayesglm <- Lrnr_pkg_SuperLearner$new("SL.bayesglm") # Bayesian GLM,
    lrn_lasso <- make_learner(Lrnr_glmnet) # Lasso Regression
    lrn_ridge <- Lrnr_glmnet$new(alpha = 0) # Ridge Regression
    lrn_enet.5 <- make_learner(Lrnr_glmnet, alpha = 0.5) # Elastic-net Penalized Regression
    lrn_polspline <- Lrnr_polspline$new() # Multivariate Adaptive Regression Splines
    Lrn_lightgbm <- Lrnr_lightgbm$new() # Gradient Boosting Machine
    Lrn_dbarts <- Lrnr_dbarts$new() # Bayesian Additive Regression Trees
    
    base_lrnrs <- c(lrn_glm, lrn_earth, lrn_ranger, lrn_bayesglm, lrn_lasso, lrn_ridge, lrn_enet.5, lrn_polspline, lrn_polspline, Lrn_lightgbm, Lrn_dbarts)
    names(base_lrnrs) <- c("GLM", "Gam", "RF", "BayesianGLM", "Lassso", "Ridge", "ElasticNet", "Splines", "XGBoost", "BayesianTrees")
    base_lrnrs_stack <- Stack$new(base_lrnrs) 
    
    sl_default <- Lrnr_sl$new(learners = base_lrnrs_stack)
    
    candidate_lrnrs <- c(sl_default, base_lrnrs) 
    names(candidate_lrnrs)[1:3] <- c("SL_defaultMeta", names(base_lrnrs))
    candidate_lrnrs_stack <- Stack$new(candidate_lrnrs)
    
  }
  
  ########################### Train discrete SL ##################################
  
  discrete <- Lrnr_cv_selector$new(loss_loglik_binomial)
  
  sl <- Lrnr_sl$new(learners = candidate_lrnrs_stack, metalearner = discrete)
  
  
  ############# how to run without parallelization: 
  set.seed(45)
  fit <- sl$train(sl3_task)
  
  ############################# summarize results ################################
  
  # lines 71 and 72 should be the same
  print(fit$coefficients)
  print(fit$fit_object$cv_meta_fit$fit_object$coef)
  
  cv_risk_table_NLL <- fit$cv_risk(eval_fun = loss_loglik_binomial)
  cv_risk_table_NLL <- cv_risk_table_NLL[cv_risk_table_NLL$learner == "Ranger", -c(1,2)] %>% 
    rename(se_nll = se, fold_sd_nll = fold_sd) %>% 
    mutate(ci_l_nll = NLL - 1.96 * se_nll,
           ci_u_nll = NLL + 1.96 * se_nll,)
  
  auc_eval_function <- custom_ROCR_risk("auc")
  cv_risk_table_AUC <- fit$cv_risk(eval_fun = auc_eval_function)
  cv_risk_table_AUC <- cv_risk_table_AUC[cv_risk_table_AUC$learner == "Ranger", -c(1,2)] %>% 
    rename(se_auc = se, fold_sd_auc = fold_sd) %>% 
    mutate(ci_l_auc = auc - 1.96 * se_auc,
           ci_u_auc = auc + 1.96 * se_auc,)
  
  # area under the precision recall curve is recommended for rare outcomes
  aucpr_eval_function <- custom_ROCR_risk("aucpr")
  cv_risk_table_AUCPR <- fit$cv_risk(eval_fun = aucpr_eval_function)
  cv_risk_table_AUCPR <- cv_risk_table_AUCPR[cv_risk_table_AUCPR$learner == "Ranger", -c(1,2)] %>% 
    rename(se_aucpr = se, fold_sd_aucpr = fold_sd) %>% 
    mutate(ci_l_aucpr = aucpr - 1.96 * se_aucpr,
           ci_u_aucpr = aucpr + 1.96 * se_aucpr,)
  
  # 
  eval_tables <- list("NLL" = cv_risk_table_NLL, "AUC" = cv_risk_table_AUC, "AUCPR" = cv_risk_table_AUCPR)
  
  if (check_VIA){
      importance_measure <- importance(fit = fit, 
                                     eval_fun = loss_loglik_binomial
                                     # type = "permute", 
                                     # covariate_groups = list("assets" = assets)
    )
  } else {
    importance_measure <- NA
  }

  
  rtns <- list("sl_fit" = fit, 
               "importance_measure" = importance_measure,
               "NLL_summary" = cv_risk_table_NLL, 
               "AUC_summary" = cv_risk_table_AUC, 
               "AUCPR_summary" = cv_risk_table_AUCPR)
  return(rtns)
}
```


```{r}
exp_var_lists <- list(c("W", "VS", "LT"),
                      c("n_W", "n_VS", "n_LT"),
                      c("n_W", "h_m_VS", "n_LT"),
                      c("n_W", "n_VS", "h_m_VS", "n_LT"),
                      c("W", "VS", "LT", "n_W", "n_VS", "n_LT"),
                      c("W", "VS", "LT", "n_W", "h_m_VS", "n_LT"),
                      c("W", "VS", "LT", "n_W", "n_VS", "h_m_VS", "n_LT")
                      )
```


```{r}
# df_sl3 <- get_df_sl(sets = exp_var_lists[[1]], prev_block = F)
# 
# # df_sl3 <- df_sl3[unique(c(1:500, which(df_sl3$Y_t_plus_1 == 1))), ]
# 
# rslts <- run_sl(df_sl3, rf_only = F)
# 
# save.image(file=here("data", "rdata", "05_VIA_sl3_all.RData"))

# knitr::purl(here("scripts", "05_VIA.Rmd"))
```


```{r}
# load(here("data", "rdata", "05_VIA_sl3_all.RData"))
```

## Random (Unifrom) Imputation
```{r}
# predicts <- list()
# var_imp <- list()
# top_picks <- list()
# vars <- c()
# 
# for(i in 1:length(exp_var_lists)){
#   cat("\n", i, ". \nIncluding", exp_var_lists[[i]], "in the predictive model: \n")
#   df_sl3 <- get_df_sl(sets = exp_var_lists[[i]],
#                    prev_block = F,
#                    inputation = "random_unif")
#   cat("     data dim:", dim(df_sl3))
#   
#   # df_sl3 <- df_sl3[unique(c(1:500, which(df_sl3$Y_t_plus_1 == 1))), ]
#   rslts_i <- run_sl(df_sl3, rf_only = T, check_VIA = T)
# 
#   predicts[[i]] = rslts_i$sl_fit$predict()
#   var_imp[[i]] = rslts_i$importance_measure
#   top_picks[[i]] <- rslts_i$importance_measure$covariate[1:10]
#   var <- paste0(exp_var_lists[[i]], collapse = ", ")
#   vars <- append(vars, var)
#   results <- data.frame(var = var) %>%
#     cbind(rslts_i$AUC_summary) %>%
#     cbind(rslts_i$AUCPR_summary) %>%
#     cbind(rslts_i$NLL_summary)
#   if(i==1){
#     results_df = results
#   } else{
#     results_df = rbind(results_df, results)
#   }
# }
# 
# save.image(file=here("data", "rdata", "05_VIA_sl3_rf_randomUnifImput.RData"))
```

### model performances
```{r}
load(here("data", "rdata", "05_VIA_sl3_rf_randomUnifImput.RData"))

# cat("All evaluations: ", names(results_df))

print(results_df %>% select(auc, ci_l_auc, ci_u_auc, aucpr, ci_l_aucpr, ci_u_aucpr, NLL, ci_l_nll, ci_u_nll))

write.csv(results_df, file = here("results", "summary_tables", "04_VIA_sl3_rf_randomUnifImpt_moldels_preformances_summary.csv"))

results_df_randomImput <- results_df
```

```{r}
df_y_y_hat <- df_sl3 |> 
  dplyr::select(ICUSTAY_ID, half_day, Y_t_plus_1,
                AGE, GENDER_F, INSURANCE_Y, LANGUAGE_ENGL, RELIGION_Chrt, MARITAL_PARTN,
                ETHNICITY_W, ETHNICITY_B, ETHNICITY_A, ETHNICITY_H, ETHNICITY_O) |>
  mutate(y_hat_values_only_randomInput = predicts[[1]],
         y_hat_patterns_only_randomInput = predicts[[4]],
         y_hat_values_patterns_randomInput = predicts[[7]])
```

## Impute with mode or median
```{r}
# predicts <- list()
# var_imp <- list()
# top_picks <- list()
# vars <- c()
# 
# for(i in 1:length(exp_var_lists)){
#   cat("\n", i, ". \nIncluding", exp_var_lists[[i]], "in the predictive model: \n")
#   df_sl3 <- get_df_sl(sets = exp_var_lists[[i]],
#                    prev_block = F,
#                    inputation = "median_mode")
#   cat("     data dim:", dim(df_sl3))
# 
#   rslts_i <- run_sl(df_sl3, rf_only = T, check_VIA = T)
# 
#   predicts[[i]] = rslts_i$sl_fit$predict()
#   var_imp[[i]] = rslts_i$importance_measure
#   top_picks[[i]] <- rslts_i$importance_measure$covariate[1:10]
#   var <- paste0(exp_var_lists[[i]], collapse = ", ")
#   vars <- append(vars, var)
#   results <- data.frame(var = var) %>%
#     cbind(rslts_i$AUC_summary) %>%
#     cbind(rslts_i$AUCPR_summary) %>%
#     cbind(rslts_i$NLL_summary)
#   if(i==1){
#     results_df = results
#   } else{
#     results_df = rbind(results_df, results)
#   }
# }
# 
# save.image(file=here("data", "rdata", "05_VIA_sl3_rf_medianModeImput.RData"))
```

### model performances
```{r}
load(here("data", "rdata", "05_VIA_sl3_rf_medianModeImput.RData"))
print(results_df %>% select(auc, ci_l_auc, ci_u_auc, aucpr, ci_l_aucpr, ci_u_aucpr, NLL, ci_l_nll, ci_u_nll))

write.csv(results_df, file = here("results", "summary_tables", "04_VIA_sl3_rf_medianModeImput_moldels_preformances_summary.csv"))

results_df_medianModeImput <- results_df
```

```{r}
df_y_y_hat <- df_y_y_hat |> 
  mutate(y_hat_values_only_medInput = predicts[[1]],
         y_hat_patterns_only_medInput = predicts[[4]],
         y_hat_values_patterns_medInput = predicts[[7]])

df_y_y_hat_long <- tidyr::pivot_longer(df_y_y_hat,
                                       cols = c("y_hat_values_only_randomInput",
                                               "y_hat_patterns_only_randomInput",
                                               "y_hat_values_patterns_randomInput",
                                               "y_hat_values_only_medInput",
                                               "y_hat_patterns_only_medInput",
                                               "y_hat_values_patterns_medInput"),
                                      names_to = "method",
                                      values_to = "y_hat") |>
  mutate(inputation = ifelse(grepl("randomInput", method), "Random", "MedianMode"),
         train_set = dplyr::case_when(grepl("values_only", method) ~ "Values solely",
                                      grepl("patterns_only", method) ~ "Patterns solely",
                                      TRUE ~ "Values and Patterns")) |>
  dplyr::select(ICUSTAY_ID, half_day, Y_t_plus_1, inputation, train_set, y_hat, everything())

```

```{r}
calculate_nll <- function(true_labels, predicted_scores) {
  epsilon <- 1e-15  # to avoid log(0)
  predicted_scores <- pmax(epsilon, pmin(1 - epsilon, predicted_scores))
  -mean(true_labels * log(predicted_scores) + (1 - true_labels) * log(1 - predicted_scores))
}

performance_age <- df_y_y_hat_long |> 
  group_by(inputation, train_set) %>%
  summarize(
    auc = as.numeric(roc(Y_t_plus_1, y_hat)$auc),
    aucpr = pr.curve(scores.class0 = y_hat[Y_t_plus_1 == 1], 
                     scores.class1 = y_hat[Y_t_plus_1 == 0], 
                     curve = TRUE)$auc.integral,
    nll = calculate_nll(Y_t_plus_1, y_hat)
  ) |>
  print()
```

# model performances by age
```{r}
performance_age <- df_y_y_hat_long |> 
  mutate(age_group = cut(AGE,
                         breaks = c(18,30,45,65,101),
                         labels = c("18-30","31-45","46-65",">65"))) |>
  group_by(age_group, inputation, train_set) %>%
  summarize(
    auc = as.numeric(roc(Y_t_plus_1, y_hat)$auc),
    aucpr = pr.curve(scores.class0 = y_hat[Y_t_plus_1 == 1], 
                     scores.class1 = y_hat[Y_t_plus_1 == 0], 
                     curve = TRUE)$auc.integral,
    nll = calculate_nll(Y_t_plus_1, y_hat)
  )

performance_age |>
  dplyr::select(age_group, inputation, train_set, auc) |>
  tidyr::pivot_wider(names_from = age_group, values_from = auc) |>
  print()

performance_age |>
  dplyr::select(age_group, inputation, train_set, aucpr) |>
  tidyr::pivot_wider(names_from = age_group, values_from = aucpr)|>
  print()

performance_age |>
  dplyr::select(age_group, inputation, train_set, nll) |>
  tidyr::pivot_wider(names_from = age_group, values_from = nll)|>
  print()
```


# model performances by race/ethnicity
```{r}
c("WHITE", "ASIAN", "BLACK", "HISPANIC", "Other", "Missing")

performance_age <- df_y_y_hat_long |> 
  mutate(ethnicity = case_when(ETHNICITY_W == 1 ~ "WHITE",
                               ETHNICITY_A == 1 ~ "ASIAN",
                               ETHNICITY_B == 1 ~ "BLACK",
                               ETHNICITY_H == 1 ~ "HISPANIC",
                               ETHNICITY_O == 1 ~ "Other",
                               TRUE ~ "Missing")) |>
  group_by(ethnicity, inputation, train_set) %>%
  summarize(
    auc = as.numeric(roc(Y_t_plus_1, y_hat)$auc),
    aucpr = pr.curve(scores.class0 = y_hat[Y_t_plus_1 == 1], 
                     scores.class1 = y_hat[Y_t_plus_1 == 0], 
                     curve = TRUE)$auc.integral,
    nll = calculate_nll(Y_t_plus_1, y_hat)
  )

performance_age |>
  dplyr::select(ethnicity, inputation, train_set, auc) |>
  tidyr::pivot_wider(names_from = ethnicity, values_from = auc) |>
  print()

performance_age |>
  dplyr::select(ethnicity, inputation, train_set, aucpr) |>
  tidyr::pivot_wider(names_from = ethnicity, values_from = aucpr)|>
  print()

performance_age |>
  dplyr::select(ethnicity, inputation, train_set, nll) |>
  tidyr::pivot_wider(names_from = ethnicity, values_from = nll)|>
  print()
```






# model performances by gender
```{r}
performance_age <- df_y_y_hat_long |> 
  mutate(gender = ifelse(GENDER_F == 1, "Female", "Male")) |>
  group_by(gender, inputation, train_set) %>%
  summarize(
    auc = as.numeric(roc(Y_t_plus_1, y_hat)$auc),
    aucpr = pr.curve(scores.class0 = y_hat[Y_t_plus_1 == 1], 
                     scores.class1 = y_hat[Y_t_plus_1 == 0], 
                     curve = TRUE)$auc.integral,
    nll = calculate_nll(Y_t_plus_1, y_hat)
  )

performance_age |>
  dplyr::select(gender, inputation, train_set, auc) |>
  tidyr::pivot_wider(names_from = gender, values_from = auc) |>
  print()

performance_age |>
  dplyr::select(gender, inputation, train_set, aucpr) |>
  tidyr::pivot_wider(names_from = gender, values_from = aucpr)|>
  print()

performance_age |>
  dplyr::select(gender, inputation, train_set, nll) |>
  tidyr::pivot_wider(names_from = gender, values_from = nll)|>
  print()
```


# model performance plots
```{r, fig.width=7, fig.height=4}
results_df <- results_df_medianModeImput %>% mutate(imputation_block = "MedianMode") %>%
              rbind(results_df_randomImput %>% mutate(imputation_block = "Random")) 


results_df$var <- factor(results_df$var, levels = rev(results_df_medianModeImput$var))
results_df$imputation_block <- factor(results_df$imputation_block, levels = c("MedianMode", "Random"))

imputation_block_color <- c("Random" = "#EEB422", "MedianMode" = "#1874CD")

results_df_tmp <- results_df[results_df$var %in% c("W, VS, LT", "n_W, n_VS, h_m_VS, n_LT", "W, VS, LT, n_W, n_VS, h_m_VS, n_LT"), ]

results_df_tmp <- results_df_tmp |>
  dplyr::mutate(var_set = case_when(
    var == "W, VS, LT" ~ "Values solely",
    var == "n_W, n_VS, h_m_VS, n_LT" ~ "Patterns solely",
    var == "W, VS, LT, n_W, n_VS, h_m_VS, n_LT" ~ "Values and Patterns",
  ))

results_df_tmp$var_set <- factor(results_df_tmp$var_set,
                                 levels = c("Values and Patterns",
                                            "Patterns solely",
                                            "Values solely"))

p <- ggplot(data=results_df_tmp, aes(x=var_set, y=auc, ymin=ci_l_auc, ymax=ci_u_auc)) +
    geom_point(shape=24, size=2, aes(col = imputation_block, fill = imputation_block)) +
    geom_errorbar(shape=24, width=0.3, aes(col = imputation_block)) +
    ylim(c(0.7,0.9)) +  
    coord_flip() +  # flip coordinates (puts labels on y axis)
    labs(x="Sets of training variables")+
    labs(y="AUC") +
    ggtitle(" ")+
    theme(plot.title = element_text(hjust = 0.5),
          axis.title.y = element_blank(),
          panel.grid.major.x = element_blank(),
          legend.position="right") +
    scale_color_manual(values = imputation_block_color) +
    scale_fill_manual(values = imputation_block_color) + 
    guides(fill=guide_legend(title="Imputation", nrow = 2),
           col=guide_legend(title="Imputation", nrow = 2))  +
  theme_bw()

p

ggsave(here("VIA_rf_AUC.png"), p, width = 6.5, height = 3, dpi = 800)

p2 <- ggplot(data=results_df_tmp, aes(x=var, y=NLL, ymin=ci_l_nll, ymax=ci_u_nll)) +
    geom_point(shape=24, size=2, aes(col = imputation_block, fill = imputation_block)) +
    geom_errorbar(shape=24, width=0.3, aes(col = imputation_block)) +
    # ylim(c(0,1)) +  # not working
    coord_flip() +  # flip coordinates (puts labels on y axis)
    labs(x="Sets of variables")+
    labs(y="Negative Log Likelihood") +
    ggtitle(" ")+
    theme(plot.title = element_text(size=8.6, hjust = 0.5),
          axis.title.y = element_blank(),
          panel.grid.major.x = element_blank(),
          axis.text.y = element_text(size=8),
          axis.title.x = element_text(size=8),
          axis.text = element_text(size=7),
          legend.position="right") +
    scale_color_manual(values = imputation_block_color) +
    scale_fill_manual(values = imputation_block_color) + 
    guides(fill=guide_legend(title="Imputation method", nrow = 1),
           col=guide_legend(title="Imputation method", nrow = 1)) 

p2


```


```{r}
results_df_tmp
print(results_df_tmp)
```


```{r}
df_rslt <- results_df_tmp[results_df_tmp$imputation_block == "MedianMode", ]

AUC <- paste0(round(df_rslt$auc, 2), " (", round(df_rslt$ci_l_auc, 2), ", ", round(df_rslt$ci_u_auc, 2), ")")
NLL <- paste0(round(df_rslt$NLL, 2), " (", round(df_rslt$ci_l_nll, 2), ", ", round(df_rslt$ci_u_nll, 2), ")")
AUCPR <- paste0(round(df_rslt$aucpr, 2), " (", round(df_rslt$ci_l_aucpr, 2), ", ", round(df_rslt$ci_u_aucpr, 2), ")")

df_rslt <- data.frame(df_rslt$var, NLL, AUC, AUCPR)
print(df_rslt)
```
